requirements.txt
pytube~=12.1.0
Pillow~=9.1.1